+++ components/main/Contact.vue
@@ -61,8 +61,8 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-  @import '../assets/styles/colors.scss';
-  @import '../assets/styles/breaks.scss';
+  @import '../../assets/styles/colors.scss';
+  @import '../../assets/styles/breaks.scss';
 
   section {
     background-color: color(blueLight);
